Remove database migration functionality and related components from the ObsidianGemini plugin. Update imports to reflect new structure and ensure compatibility with the current codebase. Refactor history management to utilize updated types.
