+++ src/history.ts
@@ -1,7 +1,7 @@
 import ObsidianGemini from '../main';
 import { Notice, TFile, debounce } from 'obsidian';
 import { GeminiDatabase } from './database';
-import { BasicGeminiConversationEntry, GeminiConversationEntry } from './database/types';
+import { BasicGeminiConversationEntry, GeminiConversationEntry } from './types/conversation';
 import { MarkdownHistory } from './markdownHistory';
 
 export class GeminiHistory {
